customize

+++ util/colors.py
@@ -10,7 +10,7 @@ def getpalette(count):
         URL_PALETTE = "http://www.colourlovers.com/api/palettes/random?format=json"
 
         response = request.Request(URL_PALETTE, headers =
-                               {"User-Agent": "Mozilla/5.0"})
+                                   {"User-Agent": "Mozilla/5.0"})
         response = request.urlopen(response)
         response = response.read()
         
@@ -46,7 +46,7 @@ def getpalette(count):
         
         data = [response["r"], response["g"], response["b"]]
         p_all.append( data )
-    
+        
     # print(palettes)
     # print(p_all)
     return p_all
@@ -61,8 +61,7 @@ def puzzleGen(rows, cols, ULC, URC, LLC, LRC):
 
         for c in range(cols):
             puzzle[r].append([])
-
-
+            
     puzzle[0][0] = ULC
     puzzle[0][cols - 1] = URC
     puzzle[rows - 1][cols - 1] = LRC
@@ -83,32 +82,32 @@ def puzzleGen(rows, cols, ULC, URC, LLC, LRC):
                            random.randint(0,255),
                            random.randint(0,255)]
     """
-            
+    
     top_change = []
     bottom_change = []
-
+    
     for i in range(3):
-        top_change.append( (puzzle[0][cols - 1][i] - puzzle[0][0][i]) / (cols - 1) )
-        bottom_change.append( (puzzle[rows - 1][cols - 1][i] - puzzle[rows - 1][0][i]) / (cols - 1) )
+        top_change.append((puzzle[0][cols - 1][i] - puzzle[0][0][i]) / (cols - 1)) 
+        bottom_change.append((puzzle[rows - 1][cols - 1][i] - puzzle[rows - 1][0][i]) / (cols - 1))
     
     for c in range(cols):
         if puzzle[0][c] == []:
             for i in range(3):
-                puzzle[0][c].append( puzzle[0][0][i] + top_change[i] * c )
+                puzzle[0][c].append(puzzle[0][0][i] + top_change[i] * c)
 
         if puzzle[rows - 1][c] == []:
             for i in range(3):
-                puzzle[rows - 1][c].append( puzzle[rows - 1][0][i] + bottom_change[i] * c )
+                puzzle[rows - 1][c].append(puzzle[rows - 1][0][i] + bottom_change[i] * c )
 
     for c in range(cols):
         change = []
 
         for i in range(3):
-            change.append( (puzzle[rows - 1][c][i] - puzzle[0][c][i]) / (rows - 1) )
+            change.append((puzzle[rows - 1][c][i] - puzzle[0][c][i]) / (rows - 1))
 
         for r in range(rows):
             if puzzle[r][c] == []:
                 for i in range(3):
-                    puzzle[r][c].append( puzzle[0][c][i] + change[i] * r)
-            
+                    puzzle[r][c].append(int(puzzle[0][c][i]) + int(change[i] * r))
+                    
     return puzzle
